% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAsearch_methods.R
\name{plot.GAsearch}
\alias{plot.GAsearch}
\title{Plot method for the GAsearch class output by \code{\link{kofnGA}}.}
\usage{
\method{plot}{GAsearch}(x, type = c("l", "l"), lty = c(1, 1),
  pch = c(-1, -1), col = c("blue", "red"), lwd = c(1, 1), ...)
}
\arguments{
\item{x}{An object of class \code{GAsearch}, as returned by \code{kofnGA}.}

\item{type}{Controls series types.}

\item{lty}{Controls line types.}

\item{pch}{Controls point markers.}

\item{col}{Controls colors.}

\item{lwd}{Controls line widths.}

\item{\dots}{Used to pass other plot-control arguments.}
}
\description{
Arguments \code{type}, \code{lty}, \code{pch}, \code{col}, \code{lwd} Can be supplied to change
the appearance of the lines produced by the plot method.  Each is a 2-vector: the first element
gives the parameter for the plot of average objective function value, and the second element
gives the parameter for the plot of the minimum objective function value. See \code{plot} or
\code{matplot} for description and possible values.
}
