% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAsearch_methods.R
\name{print.GAsearch}
\alias{print.GAsearch}
\title{Print method for the GAsearch class output by \code{\link{kofnGA}}.}
\usage{
\method{print}{GAsearch}(x, ...)
}
\arguments{
\item{x}{An object of class \code{GAsearch}, as returned by \code{kofnGA}.}

\item{\dots}{Included for consistency with generic functions.}
}
\description{
Print method for the GAsearch class output by \code{\link{kofnGA}}.
}
