% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAsearch_methods.R
\name{print.summary.GAsearch}
\alias{print.summary.GAsearch}
\title{Print method for the summary.GAsearch class used in \code{\link{kofnGA}}.}
\usage{
\method{print}{summary.GAsearch}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.GAsearch}.}

\item{\dots}{Included for consistency with generic functions.}
}
\description{
Print method for the summary.GAsearch class used in \code{\link{kofnGA}}.
}
