% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAsearch_methods.R
\name{summary.GAsearch}
\alias{summary.GAsearch}
\title{Summary method for the GAsearch class output by \code{\link{kofnGA}}.}
\usage{
\method{summary}{GAsearch}(object, ...)
}
\arguments{
\item{object}{An object of class \code{GAsearch}, as returned by \code{kofnGA}.}

\item{\dots}{Included for consistency with generic functions.}
}
\description{
Summary method for the GAsearch class output by \code{\link{kofnGA}}.
}
